Authenticated encryption.
